优化打包

+++ build-worker.js
@@ -1,9 +1,17 @@
 let path = require('path');
 let fs = require('fs');
 let workerPath = './dist/worker.min.js';
+let outputPath = '';
+
+if (process.env.NODE_ENV == 'compress') {
+    outputPath = './dist/inmap.min.js';
+} else {
+    outputPath = './dist/inmap.js';
+}
+
 let workerContent = fs.readFileSync(path.join(__dirname, workerPath), 'utf-8');
 
-fs.readFile(path.join(__dirname, './dist/inmap.min.js'), 'utf-8', function (err, data) {
+fs.readFile(path.join(__dirname, outputPath), 'utf-8', function (err, data) {
     if (err) {
         throw new Error('file not find');
     } else {
@@ -11,7 +19,7 @@ fs.readFile(path.join(__dirname, './dist/inmap.min.js'), 'utf-8', function (err,
         workerContent = workerContent.replace(/"/g, '\\"').replace(/'/g, '\\\'');
 
         let repaleData = data.replace('[workerContentString]', workerContent);
-        fs.writeFile(path.join(__dirname, './dist/inmap.min.js'), repaleData, function (err) {
+        fs.writeFile(path.join(__dirname, outputPath), repaleData, function (err) {
             if (err) {
                 throw new Error('write error');
             }
@@ -19,6 +27,8 @@ fs.readFile(path.join(__dirname, './dist/inmap.min.js'), 'utf-8', function (err,
     }
 
 });
-fs.unlink(workerPath, (err) => {
-    if (err) throw err;
-});
+if (process.env.NODE_ENV !== 'compress') {
+    fs.unlink(workerPath, (err) => {
+        if (err) throw err;
+    });
+}
